InitialServers: split by variants
