+++ app/src/main/java/page/ooooo/geoshare/data/local/database/AppDatabase.kt
@@ -2,10 +2,6 @@ package page.ooooo.geoshare.data.local.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import androidx.room.migration.Migration
-import androidx.sqlite.db.SupportSQLiteDatabase
-import kotlin.uuid.ExperimentalUuidApi
-import kotlin.uuid.Uuid
 
 @Database(
     entities = [Server::class, Link::class],
@@ -14,624 +10,4 @@ import kotlin.uuid.Uuid
 abstract class AppDatabase : RoomDatabase() {
     abstract fun getLinkDao(): LinkDao
     abstract fun getServerDao(): ServerDao
-
-    @OptIn(ExperimentalUuidApi::class)
-    companion object {
-        /**
-         * Delete all links and populate the table with initial links.
-         *
-         * When you change these links, you must:
-         *
-         * 1. Increase the version of [AppDatabase].
-         * 2. Add new migration similar to [MIGRATION_1_2].
-         * 3. Update InitialLinksTest.
-         *
-         * Optionally, you can add the new links to [page.ooooo.geoshare.data.di.defaultFakeLinks], which is used for
-         * testing.
-         *
-         * Notice that we use a standard map display link for Google Street View search link, because Google Street
-         * View doesn't support search.
-         *
-         * See https://developers.google.com/maps/documentation/urls/get-started
-         *
-         * See https://developer.apple.com/library/archive/featuredarticles/iPhoneURLScheme_Reference/MapLinks/MapLinks.html
-         *
-         * See https://web.archive.org/web/20250609044205/https://www.magicearth.com/developers/
-         */
-        fun restoreInitialLinks(db: SupportSQLiteDatabase) {
-            db.execSQL("DELETE FROM link")
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "OpenStreetMap",
-                    "OpenStreetMap",
-                    "WGS84",
-                    "DISPLAY",
-                    1,
-                    1,
-                    1,
-                    @Suppress("SpellCheckingInspection") "https://www.openstreetmap.org/?mlat={lat}&mlon={lon}#map={z}/{lat}/{lon}",
-                    "https://www.openstreetmap.org/search?query={q}",
-                    1772395295367,
-                    Uuid.parse("a771fd79-291e-4e55-9952-601f87b05bfe").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "OpenStreetMap",
-                    "OpenStreetMap navigation",
-                    "WGS84",
-                    "NAVIGATION",
-                    1,
-                    0,
-                    1,
-                    "https://www.openstreetmap.org/directions?to={lat}%2C{lon}",
-                    "https://www.openstreetmap.org/directions?to={q}",
-                    1772395295367,
-                    Uuid.parse("dad7a723-eeb1-4f60-af5d-7813b3cc1926").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Google Maps",
-                    "Google Maps",
-                    "GCJ02_MAINLAND_CHINA",
-                    "DISPLAY",
-                    1,
-                    1,
-                    1,
-                    "https://www.google.com/maps/search/?api=1&query={lat}%2C{lon}",
-                    "https://www.google.com/maps/search/?api=1&query={q}",
-                    1772395295367,
-                    Uuid.parse("7bd96da4-beba-4a30-9dbd-b437a49a1dc0").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Google Maps",
-                    "Google Maps navigation",
-                    "GCJ02_MAINLAND_CHINA",
-                    "NAVIGATION",
-                    1,
-                    0,
-                    1,
-                    "https://www.google.com/maps/dir/?api=1&destination={lat}%2C{lon}",
-                    "https://www.google.com/maps/dir/?api=1&destination={q}",
-                    1772395295367,
-                    Uuid.parse("64b0b360-24ec-4113-9056-314223c6e19a").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Google Maps",
-                    "Google Street View",
-                    "GCJ02_MAINLAND_CHINA",
-                    "STREET_VIEW",
-                    1,
-                    0,
-                    0,
-                    "https://www.google.com/maps/@?api=1&map_action=pano&viewpoint={lat}%2C{lon}",
-                    "https://www.google.com/maps/search/?api=1&query={q}",
-                    1772395295367,
-                    Uuid.parse("9d7cd113-ce01-4b8b-82fe-856956b8b20a").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Google Maps",
-                    "Google Maps Plus Code",
-                    "GCJ02_MAINLAND_CHINA",
-                    "DISPLAY",
-                    0,
-                    0,
-                    1,
-                    "https://www.google.com/maps/place/{plus_code}",
-                    "",
-                    1776849143380,
-                    Uuid.parse("5c891351-3f66-4aa2-83c4-f13c5a67fdef").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Apple Maps",
-                    "Apple Maps",
-                    "WGS84",
-                    "DISPLAY",
-                    1,
-                    0,
-                    1,
-                    "https://maps.apple.com/?ll={lat}%2C{lon}&z={z}&q={name}",
-                    "https://maps.apple.com/?q={q}",
-                    1772395295367,
-                    Uuid.parse("ce900ea1-2c5d-4641-82f3-a5429a68d603").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Apple Maps",
-                    "Apple Maps navigation",
-                    "WGS84",
-                    "NAVIGATION",
-                    1,
-                    0,
-                    1,
-                    "https://maps.apple.com/?daddr={lat}%2C{lon}",
-                    "https://maps.apple.com/?daddr={q}",
-                    1772395295367,
-                    Uuid.parse("a5092c63-cf5c-4225-9059-e888ae12e215").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Magic Earth",
-                    "Magic Earth",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    1,
-                    "magicearth://?show_on_map&lat={lat}&lon={lon}&name={name}",
-                    "magicearth://?open_search&q={q}",
-                    1772395295367,
-                    Uuid.parse("b109970a-aef8-4482-9879-52e128fd0e07").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "Magic Earth",
-                    "Magic Earth navigation",
-                    "WGS84",
-                    "NAVIGATION",
-                    0,
-                    0,
-                    1,
-                    "magicearth://?get_directions&lat={lat}&lon={lon}",
-                    "magicearth://?get_directions&q={q}",
-                    1772395295367,
-                    Uuid.parse("ee4f961c-44b0-4cb6-baad-1ed28edb8ec7").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    "GasBuddy",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    0,
-                    "https://www.gasbuddy.com/gaspricemap?lat={lat}&lng={lon}&z={z}",
-                    "",
-                    1772579164207,
-                    Uuid.parse("fd89f6f0-694e-4d96-b604-ed15e2530a2d").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    @Suppress("SpellCheckingInspection") "KartaView",
-                    "WGS84",
-                    "STREET_VIEW",
-                    0,
-                    0,
-                    0,
-                    @Suppress("SpellCheckingInspection") "https://kartaview.org/map/@{lat}%2C{lon},{z}z",
-                    "",
-                    1772579164207,
-                    Uuid.parse("7e09855d-d29b-4c18-944f-7fa440db3528").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    @Suppress("SpellCheckingInspection") "Mapilio",
-                    "WGS84",
-                    "STREET_VIEW",
-                    0,
-                    0,
-                    0,
-                    "https://mapilio.com/app?lat={lat}&lng={lon}&zoom={z}",
-                    "",
-                    1772579164207,
-                    Uuid.parse("c206d165-b2db-4030-a415-203e92cacb66").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    @Suppress("SpellCheckingInspection") "Panoramax",
-                    "WGS84",
-                    "STREET_VIEW",
-                    0,
-                    0,
-                    0,
-                    "https://panoramax.openstreetmap.fr/#background=streets&focus=map&map={z}/{lat}/{lon}&speed=250",
-                    "",
-                    1772579164207,
-                    Uuid.parse("c80ccfa6-6d22-4290-b8f5-82119f87a570").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    "PeakVisor",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    0,
-                    "https://peakvisor.com/embed?lat={lat}&lng={lon}&alt=3000&yaw=160",
-                    "",
-                    1772579164207,
-                    Uuid.parse("fe5de45b-ba0f-4d87-8ee9-979f1a701978").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    "Refuges",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    0,
-                    "https://www.refuges.info/nav?map={z}/{lon}/{lat}",
-                    "",
-                    1772579164207,
-                    Uuid.parse("1aad2356-d900-45e5-91a0-d7ee2092641d").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    "uMap",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    0,
-                    "https://umap.openstreetmap.fr/en/map/test-mapy_626288#{z}/{lat}/{lon}",
-                    "",
-                    1772579164207,
-                    Uuid.parse("94e1350a-3599-43b3-858b-59750a6f8680").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    "Zoom Earth",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    0,
-                    "https://zoom.earth/maps/satellite/#view={lat},{lon},{z}",
-                    "",
-                    1776348228136,
-                    Uuid.parse("a0b5b493-31a7-410d-9fe2-285974738ff1").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "",
-                    "Géoportail",
-                    "WGS84",
-                    "DISPLAY",
-                    0,
-                    0,
-                    0,
-                    "https://cartes.gouv.fr/explorer-les-cartes?c={lon},{lat}&z={z}",
-                    "",
-                    1778680284986,
-                    Uuid.parse("b0f1715a-6645-4ae6-a4ec-36d6e5f08c34").toByteArray(),
-                )
-            )
-        }
-
-        /**
-         * Delete all servers and populate the table with initial servers.
-         *
-         * When you change these servers, you must:
-         *
-         * 1. Increase the version of [AppDatabase].
-         * 2. Add new migration similar to [MIGRATION_1_2].
-         * 3. Update InitialLinksTest.
-         *
-         * Optionally, you can add the new links to [page.ooooo.geoshare.data.di.defaultFakeServers], which is used for
-         * testing.
-         */
-        fun restoreInitialServers(db: SupportSQLiteDatabase) {
-            // TODO Move GeoShare Server item to pro variant
-            db.execSQL("DELETE FROM server")
-            db.execSQL(
-                "INSERT INTO server (`baseUrl`,`authType`,`apiKey`,`apiKeyHeader`,`enabled`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "https://api.geoshare-app.net",
-                    "ATTESTATION",
-                    "",
-                    "",
-                    0,
-                    1779859233816,
-                    Uuid.parse("640f61e6-2bb4-41d3-9b4a-65e656564d03").toByteArray(),
-                )
-            )
-            db.execSQL(
-                "INSERT INTO server (`baseUrl`,`authType`,`apiKey`,`apiKeyHeader`,`enabled`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-                arrayOf<Any>(
-                    "https://geocode.googleapis.com",
-                    "API_KEY",
-                    "",
-                    "X-Goog-Api-Key",
-                    0,
-                    1779859252618,
-                    Uuid.parse("16b3bb06-3a3b-4853-ac06-c4bf1eb346f8").toByteArray(),
-                )
-            )
-        }
-
-        val MIGRATION_1_2 = object : Migration(1, 2) {
-            override fun migrate(db: SupportSQLiteDatabase) {
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        "GasBuddy",
-                        "WGS84",
-                        "DISPLAY",
-                        0,
-                        0,
-                        0,
-                        "https://www.gasbuddy.com/gaspricemap?lat={lat}&lng={lon}&z={z}",
-                        "",
-                        1772579164207,
-                        Uuid.parse("fd89f6f0-694e-4d96-b604-ed15e2530a2d").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        @Suppress("SpellCheckingInspection") "KartaView",
-                        "WGS84",
-                        "STREET_VIEW",
-                        0,
-                        0,
-                        0,
-                        @Suppress("SpellCheckingInspection") "https://kartaview.org/map/@{lat}%2C{lon},{z}z",
-                        "",
-                        1772579164207,
-                        Uuid.parse("7e09855d-d29b-4c18-944f-7fa440db3528").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        @Suppress("SpellCheckingInspection") "Mapilio",
-                        "WGS84",
-                        "STREET_VIEW",
-                        0,
-                        0,
-                        0,
-                        "https://mapilio.com/app?lat={lat}&lng={lon}&zoom={z}",
-                        "",
-                        1772579164207,
-                        Uuid.parse("c206d165-b2db-4030-a415-203e92cacb66").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        @Suppress("SpellCheckingInspection") "Panoramax",
-                        "WGS84",
-                        "STREET_VIEW",
-                        0,
-                        0,
-                        0,
-                        "https://panoramax.openstreetmap.fr/#background=streets&focus=map&map={z}/{lat}/{lon}&speed=250",
-                        "",
-                        1772579164207,
-                        Uuid.parse("c80ccfa6-6d22-4290-b8f5-82119f87a570").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        "PeakVisor",
-                        "WGS84",
-                        "DISPLAY",
-                        0,
-                        0,
-                        0,
-                        "https://peakvisor.com/embed?lat={lat}&lng={lon}&alt=3000&yaw=160",
-                        "",
-                        1772579164207,
-                        Uuid.parse("fe5de45b-ba0f-4d87-8ee9-979f1a701978").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        "Refuges",
-                        "WGS84",
-                        "DISPLAY",
-                        0,
-                        0,
-                        0,
-                        "https://www.refuges.info/nav?map={z}/{lon}/{lat}",
-                        "",
-                        1772579164207,
-                        Uuid.parse("1aad2356-d900-45e5-91a0-d7ee2092641d").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        "uMap",
-                        "WGS84",
-                        "DISPLAY",
-                        0,
-                        0,
-                        0,
-                        "https://umap.openstreetmap.fr/en/map/test-mapy_626288#{z}/{lat}/{lon}",
-                        "",
-                        1772579164207,
-                        Uuid.parse("94e1350a-3599-43b3-858b-59750a6f8680").toByteArray(),
-                    )
-                )
-            }
-        }
-
-        val MIGRATION_2_3 = object : Migration(2, 3) {
-            override fun migrate(db: SupportSQLiteDatabase) {
-                db.execSQL(
-                    "UPDATE Link SET srs = ? WHERE uuid = ? AND srs = ?",
-                    arrayOf<Any>(
-                        "GCJ02_MAINLAND_CHINA",
-                        Uuid.parse("7bd96da4-beba-4a30-9dbd-b437a49a1dc0").toByteArray(),
-                        "GCJ02",
-                    )
-                )
-                db.execSQL(
-                    "UPDATE Link SET srs = ? WHERE uuid = ? AND srs = ?",
-                    arrayOf<Any>(
-                        "GCJ02_MAINLAND_CHINA",
-                        Uuid.parse("64b0b360-24ec-4113-9056-314223c6e19a").toByteArray(),
-                        "GCJ02",
-                    )
-                )
-                db.execSQL(
-                    "UPDATE Link SET srs = ? WHERE uuid = ? AND srs = ?",
-                    arrayOf<Any>(
-                        "GCJ02_MAINLAND_CHINA",
-                        Uuid.parse("9d7cd113-ce01-4b8b-82fe-856956b8b20a").toByteArray(),
-                        "GCJ02",
-                    )
-                )
-            }
-        }
-
-        val MIGRATION_3_4 = object : Migration(3, 4) {
-            override fun migrate(db: SupportSQLiteDatabase) {
-                db.execSQL(
-                    "UPDATE Link SET coordsUriTemplate = ? WHERE uuid = ? AND coordsUriTemplate = ?",
-                    arrayOf<Any>(
-                        @Suppress("SpellCheckingInspection") "https://www.openstreetmap.org/?mlat={lat}&mlon={lon}#map={z}/{lat}/{lon}",
-                        Uuid.parse("a771fd79-291e-4e55-9952-601f87b05bfe").toByteArray(),
-                        "https://www.openstreetmap.org/#map={z}/{lat}/{lon}",
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        "Zoom Earth",
-                        "WGS84",
-                        "DISPLAY",
-                        0,
-                        0,
-                        0,
-                        "https://zoom.earth/maps/satellite/#view={lat},{lon},{z}",
-                        "",
-                        1776348228136,
-                        Uuid.parse("a0b5b493-31a7-410d-9fe2-285974738ff1").toByteArray(),
-                    )
-                )
-            }
-        }
-
-        val MIGRATION_4_5 = object : Migration(4, 5) {
-            override fun migrate(db: SupportSQLiteDatabase) {
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "Google Maps",
-                        "Google Maps Plus Code",
-                        "GCJ02_MAINLAND_CHINA",
-                        "DISPLAY",
-                        0,
-                        0,
-                        1,
-                        "https://www.google.com/maps/place/{plus_code}",
-                        "",
-                        1776849143380,
-                        Uuid.parse("5c891351-3f66-4aa2-83c4-f13c5a67fdef").toByteArray(),
-                    )
-                )
-            }
-        }
-
-        val MIGRATION_5_6 = object : Migration(5, 6) {
-            override fun migrate(db: SupportSQLiteDatabase) {
-                db.execSQL(
-                    "INSERT OR REPLACE INTO link (`group`,`name`,`srs`,`type`,`appEnabled`,`chipEnabled`,`sheetEnabled`,`coordsUriTemplate`,`nameUriTemplate`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "",
-                        "Géoportail",
-                        "WGS84",
-                        "DISPLAY",
-                        0,
-                        0,
-                        0,
-                        "https://cartes.gouv.fr/explorer-les-cartes?c={lon},{lat}&z={z}",
-                        "",
-                        1778680284986,
-                        Uuid.parse("b0f1715a-6645-4ae6-a4ec-36d6e5f08c34").toByteArray(),
-                    )
-                )
-            }
-        }
-
-        val MIGRATION_6_7 = object : Migration(6, 7) {
-            override fun migrate(db: SupportSQLiteDatabase) {
-                // TODO Move GeoShare Server item to pro variant
-                db.execSQL(
-                    "INSERT OR REPLACE INTO server (`baseUrl`,`authType`,`apiKey`,`apiKeyHeader`,`enabled`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "https://api.geoshare-app.net",
-                        "ATTESTATION",
-                        "",
-                        "",
-                        0,
-                        1779859233816,
-                        Uuid.parse("640f61e6-2bb4-41d3-9b4a-65e656564d03").toByteArray(),
-                    )
-                )
-                db.execSQL(
-                    "INSERT OR REPLACE INTO server (`baseUrl`,`authType`,`apiKey`,`apiKeyHeader`,`enabled`,`createdAt`,`uuid`) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-                    arrayOf<Any>(
-                        "https://geocode.googleapis.com",
-                        "API_KEY",
-                        "",
-                        "X-Goog-Api-Key",
-                        0,
-                        1779859252618,
-                        Uuid.parse("16b3bb06-3a3b-4853-ac06-c4bf1eb346f8").toByteArray(),
-                    )
-                )
-            }
-        }
-    }
 }
