+++ app/src/main/java/page/ooooo/geoshare/data/local/database/InitialData.kt
@@ -1,0 +1,9 @@
+package page.ooooo.geoshare.data.local.database
+
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
+
+interface InitialData {
+    fun restore(db: SupportSQLiteDatabase)
+    val migrations: List<Migration>
+}
