+++ app/src/main/java/page/ooooo/geoshare/data/local/database/InitialServers.kt
@@ -1,0 +1,3 @@
+package page.ooooo.geoshare.data.local.database
+
+interface InitialServers : InitialData
